% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calculate_subject}
\alias{calculate_subject}
\title{Calculate the the two ADHD RS-IV subscales for a single subject.}
\usage{
calculate_subject(item_01, item_02, item_03, item_04, item_05, item_06, item_07,
  item_08, item_09, item_10, item_11, item_12, item_13, item_14, item_15,
  item_16, item_17, item_18)
}
\arguments{
  \item{item_01}{Item #1 value.}

  \item{item_02}{Item #2 value.}

  \item{item_03}{Item #3 value.}

  \item{item_04}{Item #4 value.}

  \item{item_05}{Item #5 value.}

  \item{item_06}{Item #6 value.}

  \item{item_07}{Item #7 value.}

  \item{item_08}{Item #8 value.}

  \item{item_09}{Item #9 value.}

  \item{item_10}{Item #10 value.}

  \item{item_11}{Item #11 value.}

  \item{item_12}{Item #12 value.}

  \item{item_13}{Item #13 value.}

  \item{item_14}{Item #14 value.}

  \item{item_15}{Item #15 value.}

  \item{item_16}{Item #16 value.}

  \item{item_17}{Item #17 value.}

  \item{item_18}{Item #18 value.}
}
\description{
Calculate the the two ADHD RS-IV subscales for a single
subject.
}
\details{
The twenty items can be of class \code{numeric} or
\code{integer}.  Currently any missing item value will make
the subscale and the total score missing as well.
}

