% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calculate_sample}
\alias{calculate_sample}
\title{Calculate the the two ADHD RS-IV subscales for a single subject.}
\usage{
calculate_sample(ds_sample, map = NULL)
}
\arguments{
  \item{ds_sample}{A \code{data.frame} containing items 1
  through 18.}

  \item{map}{A 'named' \code{character} \code{vector} that
  maps the variable names in the observed dataset to the
  expected variable names (ie, \code{item_01} through
  \code{item_18}.)}
}
\description{
Calculate the the two ADHD RS-IV subscales for a single
subject.
}
\details{
The eighteen items can be of class \code{numeric} or
\code{integer}.  Currently any missing item value will make
the subscale and the total score missing as well.
}

